+++ src/main/java/org/cthing/versionparser/gem/package-info.java
@@ -20,4 +20,4 @@
 @PackageNonnullByDefault
 package org.cthing.versionparser.gem;
 
-import org.cthing.versionparser.internal.annotations.PackageNonnullByDefault;
+import org.cthing.annotations.PackageNonnullByDefault;