+++ src/main/java/org/cthing/versionparser/package-info.java
@@ -21,4 +21,4 @@
 @PackageNonnullByDefault
 package org.cthing.versionparser;
 
-import org.cthing.versionparser.internal.annotations.PackageNonnullByDefault;
+import org.cthing.annotations.PackageNonnullByDefault;